services/classifier/pyproject.toml
[project]
name = "classifier"
version = "0.1.0"
description = "HiveMind classifier service"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "nats-py (>=2.13.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
